web: sparql parser: Limit ‘tokenize-triplet-pattern’ to the triplet scope.
The ‘tokenize-triplet-pattern’ procedure would read until the end of
the query text.  So it would also try to read beyond the scope of the
triplet patterns.  By introducing the “initial-scope” mode to the
parser, we can stop it from reading beyond its scope.

This commit also adds a triplet pattern that did not end with a
“.”. This is useful for short queries that only specify a single
tripet pattern.

* web/sparql/parser.scm: Add ‘finalize-parser’ because we have
  multiple terminal points.  Add “initial-scope” mode.