web: db connections: Introduce system-wide connections.
The connections interface will change significantly with this commit.
We used to not care about where a graph's content was stored, and
importing data into a graph was outside of ‘sg-web’'s scope.

With this commit we change the connections interface so that these
shortcomings can be better addressed by the remaining codebase.

A series of follow-up commits will converge the remaining codebase to
this breaking change.

* web/www/db/connections.scm: Split connections in
  #<system-wide-connection>s and #<user-connection>s.