web: pages exploratory: Remove the example queries.
The examples queries worked for the previous version of the
Exploratory, but the current one shows only relevant information for
a single project.  It uses the system state graph to gather the
relevant items.  Users normally don't have direct access to this graph
and therefore, showing example queries becomes difficult.

* web/www/pages/exploratory.scm: Remove the example queries.