vcf2rdf: Don't process non-variants.
In multi-sample VCF files, it is common to describe variants for one
sample that are not found in another.  We used to describe variants
for every sample, even when a sample does not contain any alternative
allele for a variant.

With this commit we avoid processing variants for which no alternative
allele is found in the relevant sample.

* tools/vcf2rdf/src/vcf_variants.c: Check whether the sample has at
  least on alternative allele described in its genotype field.