vcf2rdf: Move from global sample names to local sample names.
Up until now, ‘vcf2rdf’ used a static prefix followed by the sample
name specified in the VCF file as “unique” identifier for a sample.
This worked well to “automagically” connect data based on sample name.

However, this mechanism can easily lead to linking two completely
unrelated samples with each other.  Therefore, we now use the “origin”
prefix for sample names, which is the MD5 sum of the file's contents.

* tools/vcf2rdf/include/ontology.h: Remove PREFIX_SAMPLE.
* tools/vcf2rdf/include/runtime_configuration.h: Add a bulk-alloc
  index for sample IDs.
* tools/vcf2rdf/include/vcf_header.h: Add the ‘file_hash’ parameter.
* tools/vcf2rdf/src/main.c: Pass the ‘file_hash’ to ‘process_header’.
* tools/vcf2rdf/src/ontology.c: Remove PREFIX_SAMPLE.
* tools/vcf2rdf/src/runtime_configuration.c: Add bits for the
  bulk-alloc index for sample IDs.
* tools/vcf2rdf/src/vcf_header.c: Implement generating sample IDs.
* tools/vcf2rdf/src/vcf_variants.c: Use ‘config.sample_ids’.