web: auth-manager: Implement a faster method to query Virtuoso.
Instead of using Virtuoso's HTTP interface, we use the “isql”
command-line tool to run a query.  This overcomes a little known
limitation to the HTTP interface of Virtuoso: the result set is cut
off after 10^20 rows. It should also be faster because it doesn't pass
the data through a TCP socket, but instead uses a pipe.

* web/auth-manager/api.scm: Implement the alternative way to query
  Virtuoso.
* web/auth-manager/virtuoso.scm (virtuoso-isql-query): New function.