web: Use the API for querying through the web interface.
Since the beginning of time we had this somewhat odd way of displaying
query results in the browser: We constructed a HTML table and dumped
that to the user.  Now that the API can be used to run queries, we can
use that instead.

The main benefits are twofold:
1. Less code to maintain -> all querying is done through the API.
2. The sg-web can handle large result sets because it uses streaming.

So if the user wants to render a million rows in the web browser, she
now can.

* web/Makefile.am: Remove obsolete ‘query-response.scm’.
* web/static/js/query-editor.js: Use the API to submit queries.
* web/www/pages/query-response.scm: Remove file.