web: sparql parser: Add a verbose output mode to ‘parse-query’.
I used to keep a bunch of “format” lines commented out in the parser
code to try and understand what's missing when a query fails to be
parsed.  In this commit I add those “format” statements under an
optional keyword argument.  Invoking the parser as follows will
write the debugging statements to the “current output port”:

  (parse-query ... #:debug-port (current-output-port))

* web/sparql/parser.scm: Add debugging bits.