doc: Update the “Using SPARQL with R” section.
The shift towards handling authentication through ‘sg-auth-manager’
means that the way to approach the SPARQL endpoint has changed.  On
the upside, we don't need the SPARQL package anymore, which frees us
from a few arbitrary limitations.  On the downside, we now have to
explicitely invoke a parser (while the SPARQL package would use an XML
parser).

* doc/programming-with-sparql.tex: Update text.