Introduce SGFS.
The SPARQLing-genomics FileSystem is a FUSE-based filesystem layout to
interact with a SPARQLing-genomics endpoint.

This commit mainly aims to place the proof-of-concept code into the
build tree in a non-obtrusive way.  The documentation for SGFS has
been left out because the user-facing interface is destined to change.

To build SGFS, pass ‘--with-sgfs’ to the configure script.

* configure.ac: Add bits for libfuse and SGFS.
* tools/Makefile.am: Conditionally include SGFS.
* tools/sgfs/Makefile.am: New file.
* tools/sgfs/guile/Makefile.am: New file.
* tools/sgfs/guile/sgfs/filesystem.scm: New file.
* tools/sgfs/src/sgfs.c.in: New file.