web: sparql parser: Fix bug in the parser related to FILTER statements.
The bug can be triggered by adding more than 3 items to a FILTER
statement:
  FILTER (?s = "A" OR ?s = "B" OR ?s "C" OR ?s "D")

The quotes would get added to as tokens upon entering whitespace.
After three tokens are added, it would get processed as a triplet.

This commit makes sure the bogus tokens aren't added by adding a check
to the whitespace path that checks whether the parser is in black-mode.

* web/sparql/parser.scm: Add black-mode check to the whitespace
  codepath of ‘tokenize-triplet-pattern’.