web: sparql stream: Fix bug in ‘direct-stream’.
In 0e8a4bd4916ca08b30281f52d49f2967070dad61 we introduced a parameter
to pass a custom delimiter to the stream functions.  We implemented
this for all but the ‘direct-stream’ function, because that function
doesn't use it.

However, this caused a strange interface that leads to not being able
to call all of the streaming functions with the same parameters.

This commit straightens the interface and makes ‘direct-stream’ take
the delimiter argument and do nothing with it.

* web/sparql/stream.scm: Smoothen the interface of the streaming
functions.