web: sparql parser: Simplify handling of comments.
This commit removes the separate step to remove comments before
parsing a query.  Doing so enables us to better determine whether a
hash sign must be treated as a comment or not.  It also cleans up the
code quite nicely.

It solves the bugs where a full URI with a hash sign in the triplet.

* web/sparql/parser.scm: Deal with comments inside the parser.