tools: vcf2rdf: Add option to include non-variants.
In a8462c637a54f92ef9d7ced1fb173fabc6484b7a I implemented an
optimization to not process the variants that do not defer from the
reference.  This allowed large multi-sample VCFs to be described in
significantly less space and time.

This commit adds an option to disable this optimization.

* tools/vcf2rdf/include/runtime_configuration.h: Add configuration
  option to keep non-variants.
* tools/vcf2rdf/src/runtime_configuration.c: Likewise.
* tools/vcf2rdf/src/ui.c: Add the UI option to include non-variants.
* tools/vcf2rdf/src/vcf_variants.c: Implement the inclusion option for
  non-variants.