web: project-graphs: Bugfix for the remove member button when using ORCID.
Before this commit, the full name of the user was used as identifier
to remove a member of a project.  This worked for all cases except
when using ORCID as identity provider.  This commit fixes this bug by
adding an extra “id” value to the output of ‘project-members’ that
will always contain the identifier to remove a member.

* web/www/components/project-graphs.scm: Use the “id” column to remove
  a member.
* web/www/db/projects.scm (project-members): Add extra “id” column to
  the output.